some minor cleanup and clangfmt

+++ kernel/src/tss.c
@@ -1,7 +1,8 @@
+#include "../includes/tss.h"
 
-#include "tss.h"
-
-tss_entry kernel_tss = { // I believe that these should be set before a context switch to have the most up to date values
+tss_entry kernel_tss = {
+    // I believe that these should be set before a context switch to have the
+    // most up to date values
     0, // link;
     0, // esp0
     0, // ss0; //stored stack segs go on lower 2 bytes
@@ -18,28 +19,24 @@ tss_entry kernel_tss = { // I believe that these should be set before a context 
     0, // ebx;
     0, // esp;
     0, // ebp;
-    0, // esi; 
-    0, // edi; 
+    0, // esi;
+    0, // edi;
     0, // es; //segments go on the lower 2 bytes
     0, // cs;
     0, // ss;
     0, // ds;
     0, // fs;
-    0, // gs; 
+    0, // gs;
     0, // ldtr;
     0, // iopb; // iopb goes on the higher 2 bytes
     0, // ssp;
 };
 
-
 void setup_tss() {
-
-    GDTEntry* tss_entry = get_entry(TSS_IDX);
-    set_gdt_entry(tss_entry, &kernel_tss, sizeof(kernel_tss) - 1, tss_entry->access, tss_entry->flags);
-    printf("value is: %p\n", &kernel_tss);
-    tss_descriptor kernel_tss_desc = {
-        TSS_OFFSET,
-        *tss_entry 
-    };
-    init_tss(&kernel_tss_desc);
+  GDTEntry *tss_entry = get_entry(TSS_IDX);
+  set_gdt_entry(tss_entry, (uint32_t)&kernel_tss, sizeof(kernel_tss) - 1,
+                tss_entry->access, tss_entry->flags);
+  printf("value is: %p\n", &kernel_tss);
+  tss_descriptor kernel_tss_desc = {TSS_OFFSET, *tss_entry};
+  init_tss(&kernel_tss_desc);
 }
